dashboard-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
redis==5.0.1
kafka-python==2.0.2
httpx==0.25.2
reportlab==4.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
grpcio==1.60.0
grpcio-tools==1.60.0
